% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originr-package.R
\docType{data}
\name{nsr_countries}
\alias{nsr_countries}
\title{Vector of country names for use with NSR}
\format{A vector of countries of length 251}
\description{
Vector of country names for use with NSR
}
\keyword{data}
