% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originr-package.R
\docType{package}
\name{originr-package}
\alias{originr}
\alias{originr-package}
\title{originr - Species Origin Data}
\description{
originr - Species Origin Data
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

